refactor: add factory method to create config from path
